Tweaks to regeneration functions. Ability to read matrices directory.

+++ run_one/action_handlers/SleepHandler.py
@@ -7,4 +7,4 @@ from run_one.util.util import Action
 class SleepHandler(AbstractActionHandler):
 
     def process(self, action: Action):
-        time.sleep(int(action.extra_data.get('Time', 0)) / 1000)
+        time.sleep(float(action.extra_data.get('Time', 0)) / 1000)
